requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
python-multipart==0.0.6
google-generativeai==0.3.1
pydantic==2.4.2
uuid==1.30
httpx==0.25.1
pytest==7.4.3
beautifulsoup4==4.12.2
markupsafe==2.1.3 